lab5/stock_crew/pyproject.toml
[project]
name = "stock-crew"
version = "0.1.0"
description = "NCU Agentic AI Lab5 - CrewAI multi-agent stock analysis"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "yfinance>=0.2.50",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "pandas>=2.0.0",
    "weasyprint>=68.1",
    "markdown>=3.10.2",
]
